Renamed folders and reorganized file structure
