+++ src/Components/Header/Header.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import './header.css'
+import './Header.css'
 
 export default function Header() {
   return (
